Refs #36 - Rename 'objects' to 'classes' and move most things under it.
