+++ application/src/classes/abstract-marker.js
