+++ application/src/classes/obstacles/abstract-dynamic-obstacle.js
