+++ application/src/classes/obstacles/abstract-static-obstacle.js
