+++ application/src/classes/obstacles/aspirin-bottle.js
