+++ application/src/classes/obstacles/aspirin-pill.js
