+++ application/src/classes/obstacles/clown-nose.js
