+++ application/src/classes/obstacles/comb.js
