+++ application/src/classes/obstacles/dynamic-box.js
