+++ application/src/classes/obstacles/floss.js
