+++ application/src/classes/obstacles/obstacle-factory.js
