+++ application/src/classes/obstacles/razor.js
