+++ application/src/classes/obstacles/static-box.js
