+++ application/src/classes/obstacles/toothbrush.js
