+++ application/src/classes/states/examples/car-driving-state.js
@@ -1,9 +1,8 @@
 'use strict';
 
 var Phaser          = require('phaser');
-var _               = require('underscore');
-var CarFactory      = require('../objects/car-factory');
-var ObstacleFactory = require('../objects/obstacles/obstacle-factory');
+var CarFactory      = require('../../car-factory');
+var ObstacleFactory = require('../../obstacles/obstacle-factory');
 
 var makeObstacles = function(context) {
     for (var i = 0; i < 10; i += 1) {
