+++ application/src/classes/states/menus/main-menu-state.js
@@ -6,9 +6,9 @@ var DrawImageState   = require('../examples/draw-image-state');
 var CarDrivingState  = require('../examples/car-driving-state');
 var TrackMarkerState = require('../examples/track-marker-state');
 var BathroomObstacleSetState = require('../examples/bathroom-obstacle-set-state');
-var TrackLoaderState = require('../states/track-loader');
-var TrackLoader      = require('../objects/track-loader');
-var trackList        = require('../../assets/tilemaps/maps/list');
+var TrackLoaderState = require('../track-loader');
+var TrackLoader      = require('../../track-loader');
+var trackList        = require('../../../../assets/tilemaps/maps/list');
 var _                = require('underscore');
 
 var MainMenuState = function()
