+++ application/src/classes/track-marker-factory.js
